Replace python2 with python3 in d/tests/control. (Closes: #936952)
At the time of writing, file(1) in unstable could not detect byte-compiled
files produced by python3. (It worked for python3 version 3.7 but unstable
has version 3.8.) It rendered the tag source-contains-prebuilt-python-object
untestable.

For the time being, two fictitious tags were added to the affected test
according to a best estimate. In one uncertaintly, the offending file names
incorporate the Python3 version.

The test was also marked TODO with an explanation and some instructions. The
test suite expects the test to fail until the Todo field is removed.

The test should be adjusted when Bug#950516, which for some time blocked the
bug taken care of here, was fixed.