Reduce visibility level of redundant-globbing-patterns to pedantic. (Closes: #967961)
As the bug reporter pointed out, policy is silent on redundant globbing
patterns in a single Files field. Some people may find them confusing and
perhaps split such fields, but that is a matter of style and not an error.
The tag was probably miscategorized when it was first introduced.

Reduces the severity to pedantic, the lowest level available.