Drop Lintian version from tag URLs offered in standalone HTML mode.
With the popularity of the Lintian job in the standard Salsa CI pipeline
expected to skyrocket, these links should be functional as soon as possible.

The Lintian version was added to the URL to provide greater accuracy as tags
change, but only a few versions are in use at a particular time.

For the standalone HTML mode, the version from unstable will run almost
exclusively. Those tags can usually be found on lintian.d.o. Also, tags do
not change frequently enough to make it a huge issue.

Restores a measure of usability to the vast majority of the links on that
particular page, which have been broken since they were first provided.