Set default Data vendor in Profile for early access while initializing tags.
The manual references, which are currently a Lintian::Data source,  are
required to generate descriptions while Profiles are being loaded.

This initializes the default Data vendor in Profiles and before Profiles are
fully loaded to enable early access to the manual references while the tags
are being loaded.

Gbp-Dch: ignore