Use Text::Markdown::Discount to facilitate Markdown content in tag explanations.
The explanations in tag descriptions already contain HTML. The processing
took two different forms. For HTML output, paragraphs were framed manually
with <p></p>. For plain text output, hyperlinks were removed for clarity.

For HTML, this removes the custom processing. Runs the text through a
markdown converter instead. That also frames the stub HTML and, as an
additional benefit, permits the use of Markdown annotations.

We plan to convert as many of the HTML tags in explanations to Markdown as
possible. The tag explanations can and will continue to contain HTML tags
and entities.

For plain text processing, more filters were added to improve the clarity of
the resulting text for readers. For example, Markdown annotations for
hyperlinks are being removed.

The plain text processing will be improved further as Markdown annotations
are added to the explanations. Many of them will be filtered out for plain
text output.

The Perl module Text::Markdown::Discount was chosen because it is actively
maintained. It uses a C library and is super fast (but the article is old):

    http://neilb.org/reviews/markdown.html#performance

The Discount C library and its extensions to the Markdown standard are
documented here:

    https://www.pell.portland.or.us/~orc/Code/discount/

Adds libtext-markdown-discount-perl to Depends in d/control.

The HTML output is no longer reformatted with Text::Wrap to make the page
source more legible.

In the future, we may try to use a tool like prettier [1] instead.

[1] https://prettier.io/docs/en/index.html