Replace plain to URL bug references with Bug#... for consistency.
Any of these would have been annotated with a hyperlink when listed as
reference information. This is merely for consistency.

In contexts such as git commit messages, leading hash symbols are sometimes
interpreted as comments that should be removed. It becomes an issue when
copying and pasting without thinking about the potential consequences.

The choice of the prefixed alternative as the default for Lintian was meant
as a safeguard against such troubles. The BTS appears to do the same thing.

Gbp-Dch: ignore