Generate HTML output in Output::HTML; drop; drop color HTML generation in Output::EWI.
Ties lintian-info to an output engine and will, after that program was
split, allow for multiple output formats.

Moves HTML-related routines to better places.

Also drops the awkward generation of color HTML in Output::EWI. It serves no
apparent purpose If needed, it should be reimplemented as an 'HTML fragment'
output mode (as opposed to the existing 'standalone HTML' output mode).

Gbp-Dch: ignore