Split bin/lintian-info into annotate-lintian-hints and explain-lintian-tags.
The purpose of the split is to eventually accommodate additional output
formats in explain-lintian-tags, for example for JSON. They are needed to
seed lintian.d.o with tag definitions for each new Lintian version.

For installed packages, a link is further being set from /bin/lintian-info
to annotate-lintian-hints, as that seems to be the dominant use.