In calls to IPC::Run3::run3, use /dev/null instead of redirecting handles from parent process.
The backslash before 'undef' seems like a small addition, but could make a
big difference. It indicates that the respective input or output should be
taken from or given to /dev/null instead of the corresponding handle from
the parent process.

Unfortunately, it does not solve the issues with 'man' stalling in the check
documentation/manual on some Asian manual pages, for example in manpages-zh.

Gbp-Dch: ignore