Widen output when looking for groff errors in manual pages. (See: #892423)
In archive-wide runs, Lintian hung and slowly used up all of a machine's
kernel I/O buffers when processing sme Asian-language manual pages. The
problem went away when MANWIDTH was increased.

More information may be available at the bug above. This is a triage commit.
It should be reverted when the issue was fixed in groff.

Why did this remain undetected in Lintian for so long? Perhaps the switch
from IO::Async to IPC::Run3 somehow revealed the bug. Maybe the command did
not always run before. It is not clear.

The problem occurred in the following binary packages:

    apt_2.1.10_amd64.deb
    dos2unix_7.4.1-1_amd64.deb
    manpages-zh_1.6.3.4-1_all.deb
    mplayer_1.3.0-8+b9_amd64.deb
    wesnoth-1.14-core_1.14.13-1_amd64.deb

Here is a small reproducer, courtesy of Jakub Wilk:

    $ mkdir -p usr/share/man/ja/man5
    $ printf '\343\201\210\343\201\260\343\200\201.lcs.mit.edu_debian_dists_unstable_contrib_binary\\-i386_Release            \n' > usr/share/man/ja/man5/test.5
    $ MANWIDTH=80 man --warnings -E UTF-8 -l -Tutf8 -Z usr/share/man/ja/man5/test.5 > /dev/null
    troff: <standard input>:1: warning [p 1, 0.0i]: cannot adjust line
    [hangs indefinitely]

Thanks to Jakub Wilk for tracking down the problem in groff and for
immediately suggesting a fix!