Move number-of-patches to separate check; fix autopkgtest and perltidy; rename test.
The previous two commit were improved in a few ways. The master branch
should now pass the autopkgtest job on Salsa CI. The commit from the MR was
also not tidy.

In addition, the following functional changes were made:

The tag was moved to its own check. We are trying to disentangle large
checks. The count is unrelated to other tags about patches.

The method Index::Item->unpacked_path is usually not needed anymore. The
same module now provides a more convenient way to get the Unicode content of
files. (The MR read the raw byte content of the series file.)

The test was renamed to 'two-patches'. We no longer name tests after checks
or tags. The name should describe the test package, and not its intent. For
example, there could be other test packages for the new tag with three or
four Debian patches.

The tag does not look at installation package, so a source skeleton was
sufficient for the test. It builds faster.

Also, the test package is about Debian patches that are not customarily used
in native packages. The test package was converted to a non-native skeleton.
That in turn revealed that no patch files were included, and unpacking of
the source failed. Empty dummy patches were added to satisfy dpkg-source.

On the tag name, number-of-patches is clear and friendly, but patch-count
would have been shorter. The name could also more specific in that it refers
to the number of patches applied and not to the number of patches shipped.

As for the tag context, we may eventually like to include the name of the
series file. Some people use vendor specific series files (see #922531),
although their use is prohibited in Debian by policy 4.17. The name leaves
no doubt as which count is given.

Regenerates the profiles to include the new check.

Thanks to Peter Wienemann for this valuable contribution. We look forward to
seeing this new tag in action!