Remove last vestiges of run_cmd and do_fork from Lintian::Util and beyond.
The custom signal settings in Lintian::Util were removed as a precaution
while tracking down an odd interaction with xmonad. Under that window
manager, the test suite does not ignore SIGWINCH and errors out with:

    Can't use an undefined value as an ARRAY reference at /usr/share/perl/5.28/TAP/Parser/Multiplexer.pm line 138.

Half the interaction was tracked down to that TAP module, in which a call to
select() is interrupted. Forcing another there read remains ineffective,
however, unless the SIGWINCH handler is set to IGNORE in ./private/runtests.
The investigation is ongoing.

The effort in this commit potentially revealed the source of the haunting
Heisenbug. The check debian/po-debconf made a call to run_cmd and do_fork,
two routines in Lintian::Util that were thought to be retired. Those
routines are now gone for good.

Tests for the ./commands folder were disabled so that the code there could
be kept around until it is fully implemented in the new website. Missing
routines caused otherwise unsurmountable errors.

Gbp-Dch: ignore