Stop linking input files to pool directory; issue tar errors in Processable.
The direct links are now longer needed because the information is available
elsewhere. This does not yet stop links for other files, such as files
mentioned in a *.dsc source specification.

Separates the checks further from the services offered by Processable. Drops
the checks for tar errors and and issues those tags in Processable.

The tags for unpack errors were also renamed and generalized. Other
utilities besides tar are used to unpack input files.

Gbp-Dch: ignore