Fix a broken link in the manual pages.
The link was not properly changed in recent commits when lintian-info was
split into two program.

In our updated Lintian terminology, 'tags' describe conditions that should
be fixed in general terms. Lintian then adds a situation-specific context
for the maintainer's benefit. The combined message, which is specific to a
package, is called a hint.

The manual page confused the two terms.

Gbp-Dch: ignore