Repurpose diff-contains tags to work on all Debian changes in other formats.
Reduces the number of tags issued for nuisance files in upstream sources,
but sharply increases the issuance for such files when a maintainer has
control over them.

Until now, most checks for nuisance files were turned off for formats other
than 1.0. According to trends.d.n, however, over ninety percent of packages
in unstable use format 3.0 (quilt).

As a result of a change made recently, the orig index is now also unpacked.
It is thus possible to compare those files with the corresponding sources in
the 'patched' index, which are used to build packages.

By virtue of that comparion a very detailed and accurate picture emerges as
to the maintainer's impact on the source tree. Substantial benefits are
expected, especially for non-native packages, which are extremely common in
the archive.

Renames some tags named for the 1.0 source format from 'diff-contains-*' to
a format-neutral 'debian-adds-*'.

The cruft check will be split into smaller pieces in the near future.