Prepare parsing of readelf output for xargs-style processing.
This pre-parser should allow the use of xargs-like strategies for the calls
to readelf, either directly via xargs or by using a similar command grouping
strategy implemented Perl.

The code explains it thus:

    Special case - readelf will not prefix the output with "File:
    $name" if it only gets one ELF file argument, so act as if it did...
    (but it does "the right thing" if passed a static lib >.>)

    In fact, if readelf always emitted that File: header, we could
    simply use xargs directly on readelf and just parse its output
    in the loop below.

A Perl implementation of xargs would know when readelf receives only one
argument, and could compensate for the missing line with a File indicator,
but such acrobatics are probably not necessary with the change here.

Gbp-Dch: ignore