Make objdump_info an attribute of Processable; drop last remnant of Unpacker.
Simplifies initialization of Processables. They are now unpacked on demand
the first time a feature is used.

The new architecture finally replaces in full the functionality originally
provided by the 'Needs-Info' annotations in the POD the Unpacker needed
years ago.

The collection of information is now triggered by the checks as needed, but
without explicit specifications that are difficult to maintain and likely to
provoke human error.

When only some checks are selected via the command line, only the
information needed for those checks is collected.

Gbp-Dch: ignore