Remove obsolete scripts-missing-versioned-depends-on-init-system-helpers-without. (Closes: #932378)
Based on the version of init-system-helpers that is now available in stable,
I saw no way to remedy or otherwise repurpose the tag. It is obsolete.

This commit drops the tag from Lintian.

From the original bug report:

> init-system-helpers version in stable are 1.56+nmu1.
> This makes the required versioning superfluous.

Thanks to Jörg Frings-Fürst <debian@jff.email> for the original report (made
over a year ago), for his patience, and also for the ping earlier today.