Fold Index subclasses into their Processable attributes and use the Index base class instead.
Simplifies greatly the logic behind the initialization of file indices. It
also finally unmantles, with small vestiges in Processable::Installable, the
logic of the old Unpacker.

File indices are now unpacked an initialized when they are used for the
first time. The relevant code can be located easily in the Processable
attribute that provides the respective index.

A basic example can be found in Processable::Installed, while the version in
Processable::Patched is more advanced.

A key element of the new data flow is that tags can be issued anywhere that
a Processable is available. In the present case, tags are issued when there
are errors during unpacking.

In the old Unpacker, the tags were issued at later point in checks based on
temporary files. That separation was both brittle and confusing, which is
why the new data flow is superior.

Gbp-Dch: ignore