Enforce code style as part of autopkgtest.
Many of the third-party contributions we receive are only vetted in Salsa,
but the autopkgtest job does not enforce the same code style the maintainers
are subject to when running ./private/runtests locally.

As a result, we frequently have to make follow-on commits to tidy up code.

For example, the previous commit was not perltidy-clean. This commit should
cause the pipeline to fail.

This commit enables a variety of code style tests (aka 'the scripts') as
part of autopkgtest. They were disabled for good reason, namely to avoid
unnecessary build failures. The decision was not made lightly. Complications
are possible.

One complication may be that perltidy's idea of 'tidy' changes unexpectedly
over time, and can differ between Debian releases.

Fortunately, our tests contains additional safeguards that prevent style
tests from running when installed perltidy and perlcritic versions do not
match minimum versions known to agree with the style in our repository.
Backports should not be affected.

Future issues with new version of perltidy or perlcritic should be addressed
with two commits: One to make the code tidy, and another to bump the minimum
version of those tools required to pass the tests.

The merits of this commit are presently unclear. Fixups are expected.