Install .perlcriticrc into /usr/share/lintian; matches code with specifications.
Out of all possible consequences, this was the one I least expected. After
some reflection, it made sense to keep the standard for code style, namely
the perlcritic selection, near the code being evaluated.

If the failures prove too difficult to resolve, an alternative might be to
disable the style tests again (i.e. revert commit acd41709). Instead, the
tests could be a separate job in the Salsa CI pipeline.

Credit for that latest idea belongs to Antonio Terceiro. Thank you!

Gbp-Dch: ignore