Remove two outdated harness tests for old check specifications; may restore CI pipeline.
The older check specifications parsed by some tests scripts have not been
used for a while. They were not modular and are now obsolete. Removes the
affected scripts.

There is some hope this commit may restore the Salsa CI pipeline, but the
failed log from Salsa also shows an untested tag:

    Unknown/missing tag: inconsistency-debian-watch

It is a Javascript team tag that was not imported in commit 8a665222. Our
sources do not contain the string. It is a mystery how the tag (or the
string) appeared in the failed Salsa log.

Our test suite probably located the tag via pkg-js-tools, which was
installed as a test prerequisite via d/tests/control.

Gbp-Dch: ignore