Do not install pkg-js-tools as part of autopkgtest until path conflicts are resolved.
Several tests require a debhelper build system that ships in pkg-js-tools.
That package also installs a custom tag. Lintian cannot tell the tag is from
a third party and fails its autopkgtests because the tag is untested.

Lintian does not presently offer ways to test third-party tags. Such a
facility is also unlikely to be provided in the near future.

A good long-term solution would be to install third-party tags in a
different location, such as usr/share/lintian/contrib/${package}/tags.

The package pkg-js-tools is the only known package with a custom tag. This
commit prevents it from being installed in autopkgtests until the
installation paths are adjusted.

As an aside, the tag inconsistency-debian-watch may at some point be
integrated into mainline Lintian like other tags from that package but the
check for it uses the network, which would be a controversial change for
Lintian.

Either way, the latter would not be a permanent solution for the issue.