Reduce read buffer size when duplicating tar stream. (Closes: #971895)
The read buffers were variously dimensioned before the four file indices
(for installed, control, patched and orig files) were unified in the same
code. Out of the sizes used, the idea of 4096 "tar records" seemed sensible
but causes a hanging program and zombie processes in older kernels (< 5.7).

For details, please see Bug#971895.

This commit changes the buffer size uniformly to 4K, which the bug reporter
stated works better. It is unclear if this change will also improve some
timeouts in Lintian jobs occurring on Salsa CI, such as:

    https://salsa.debian.org/debian-hamradio-team/svxlink/-/jobs/1030952

Thanks to Baptiste Beauplat for raising this issue, and for finding a fix!