Remove obsolete documentation about various 'desc' files.
This part of the documentation contains wildly inaccurate information.
It has not been updated in years.

More significant, the specifications also change frequently at the
moment so that updating the file is probably not practicable right
now.

In several cases, such as tags and build specifications in tests, the
files are not even called 'desc' anymore.

Gbp-Dch: ignore