Fix prefix detection for single orig components with only one directory component.
The algorithm to detect and remove unwanted common prefixes from paths
in orig indices is currently not working well. Usually, the problems
arise with sources containing multiple orig components, but that did
not explain a related-looking message in

    http://archive.kali.org/kali/pool/main/c/covenant-kbx/covenant-kbx_0.5-0kali1.dsc

The message was:

    Warning in group covenant-kbx/0.5-0kali1: Not exactly one top-level child at /lcl/lechner/lintian/git/bin/../lib/Lintian/Index.pm line 491.
    warning: cannot run cruft check on package source:covenant-kbx/0.5-0kali1
    skipping check of source:covenant-kbx/0.5-0kali1

This was a genuine bug in the common prefix detection. It disregarded
all plain files in the top level of the tarfile and erroneously
detected a common path component when otherwise only a single
directory was present. (The detected common prefix was the name of
that directory.)

This commmit removes the offending message from the dsc mentioned
above, and hopefully cures the bug without introducing any others.

Thanks to Raphaël Hertzog for bringing this to our attention!