Mark testsuite-autopkgtest-missing as old name for missing-tests-control. (Closes: #972844)
This fixes one third of the reported bug. With this commit, the hint

    E: musescore2 source: malformed-override Unknown tag testsuite-autopkgtest-missing in line 5

use the rename detection to instead note, more correctly:

    I: musescore2 source: unused-override missing-tests-control
    P: musescore2 source: renamed-tag testsuite-autopkgtest-missing => missing-tests-control in line 5

At the same time, d/tests/control is not present in the sources, so
more investigation is needed for the 'unused-override' message. That
will be addressed in Bug#972875.

Thanks to Thorsten Glaser for reporting these issues!