Drop unnecessary Python references from test maintainer-scripts/generated/scripts-control-interpreters. (See: !334)
This test was duplicated when the test suite was broken up into
smaller, unit-like pieces, but many redundant references were never
removed. This takes care of Python references that do not affect the
test results for the check this test is exercising.

As background, the script interpreter /usr/bin/python is going away in
Bullseye, and possibly even sooner in unstable. This prepares Lintian
for the dropping of unversioned Python interpreters in checks and
tests.

This commit is part of a series originally submitted and discussed in
the following merge request:

    https://salsa.debian.org/lintian/lintian/-/merge_requests/334

Gbp-Dch: ignore