Update two unversioned Python references in two WAF-related tests (See: !334)
As background, the script interpreter /usr/bin/python is going away in
Bullseye, and possibly even sooner in unstable. This prepares Lintian
for the dropping of unversioned Python interpreters in checks and
tests.

This commit is part of a series originally submitted and discussed in
the following merge request:

    https://salsa.debian.org/lintian/lintian/-/merge_requests/334

Gbp-Dch: ignore