Drop unversioned Python references from check 'languages/python'.
The script interpreter /usr/bin/python is going away in Bullseye, and
possibly even sooner in unstable.

Drops unversioned Python interpreters from Lintian's checks, but does
not yet flag their use as a packaging error. Users will instead see a
more generic message similar to:

    W: unusual-interpreter files/feedparser.py #!python

This commit is part of a series originally submitted and discussed in
the following merge request:

    https://salsa.debian.org/lintian/lintian/-/merge_requests/334

Thanks to Louis-Philippe Véronneau for a draft of this change!