Accept postgresql-all as an additional build prerequisite for the pgxs debhelper addon. (Closes: #973503)
The additional build prerequisite was added in all four places where
postgresl-server-dev-all was already listed.

In addition to the regular 'pgxs' debhelper addon, the installable
package postgresql-all now also satisfies the use of the 'pgxs_loop'
addon and the two pgxs-related debhelper tools:

    dh_make_pgxs
    dh_pgxs_test

Please reopen the bug if that is not what you wanted.

Thanks to Christoph Berg for bringing this issue to our attention!