Make JSON output reproducible; remove explicit maintainer field there.
The job start and end times are better handled in calling
infrastructure, which also collects output to stderr.

Thanks to David Bremner for the general observation that the output
was not reproducible!

Furthermore removes from JSON the maintainer information, which is
likewise better determined in the calling infrastructure. One reason
is that the information could vary between the different input files
within a group.

The maintainer is still available from the classification tag
'trimmed-field' when the context starts with "Maintainer".