No longer cache file content as bytes and decoded Unicode in Index::Item.
Any examination of text files encoded in UTF-8 brings a number of
inefficiencies. In the past, caching some of the content seemed like a
reasonable compromise between speed and memory use. This commit is an
attempt to reduce that memory usage.

In archive-wide runs, the two packages firefox and firefox-esr now
take up about 10G of memory apiece. Each package contains about
300,000 files of source code and thus presents an extraordiary strain
on any mechanism that deals with them.

This commit is an attempt to ease the burden on memory requirements in
anticipation of automated deployment.

Memory usage is being monitored elsewhere. There still is disk caching
at a lower level. Smarter, adaptive cache implementations are also
possible, if this commit makes Lintian too slow.

These changes may be reversed if they do not bring any anticipated
benefits.

Gbp-Dch: ignore