Accept Build-Depends-Packages (plural) as a valid field in symbols control files. (Closes: #974971)
Prevents the unknown-meta-field-in-symbols-file from being issued.

There appear to be no packages with the plural field in the archive,
perhaps because the tag would prevent them from being accepted.
Instead, the resolution of this request was tested against the
non-public packages mentioned in the filing.

The following tag disappears as a result of this commit:

lechner@lechner-desktop /l/l/l/git> bin/lintian libgdk-pixbuf-2.0-0_2.40.0+dfsg-6_amd64.deb
E: libgdk-pixbuf-2.0-0: unknown-meta-field-in-symbols-file Build-Depends-Packages, line 3