Explicitly create root Items for totally empty Indices; drop allow_empty.
The allow_empty option was never used and made no sense because root
entries are needed for any tpe of resolution.

Items for missing intermediate directories were already created on the
fly. Adds an explicit creation of root entries. They are needed for
completely empty orig tarballs, such as in suckless-tools_45-1.dsc.

It eliminates the following type of error:

    Warning in group suckless-tools/45-1: Can't call method "children" on an undefined value at /lcl/lechner/taxiv/lintian/bin/../lib/Lintian/Index/Control/Scripts.pm line 61.
    warning: cannot run cruft check on package source:suckless-tools/45-1
    skipping check of source:suckless-tools/45-1

Solves the same problem in the following sources:

main:
    roundcube-plugins-extra_1.4.8+1-1.dsc

non_free:
    dwarf-fortress_0.47.04-1.dsc
    nvidia-cg-toolkit_3.1.0013-5.dsc
    scribus-ng-doc_1.5.5+dfsg-1+transition.dsc