Collapse successive slashes in file paths extracted from tar.
Some tarballs can contain multiple slashes where one would have been
sufficient. Slashes cannot appear in file names, so it is safe to
replace with single occurrences.

Doing so eliminates error messages, especially for older tarballs.  An
example is emboss-explorer_2.2.0-10.dsc, which is fourteen years old.

The following messages disappear as a result of this commit:

    lechner@lechner-desktop /l/l/l/git> bin/lintian /mirror/debian/pool/main/e/emboss-explorer/emboss-explorer_2.2.0-10.dsc
    Warning in group emboss-explorer/2.2.0-10: Common path segment has no length at /lcl/lechner/lintian/git/bin/../lib/Lintian/Index.pm line 575.
    warning: cannot run cruft check on package source:emboss-explorer/2.2.0-10
    skipping check of source:emboss-explorer/2.2.0-10