Only reduce consecutive slashes in path names and hardlink targets.
For symbolic link targets, we have the tag symlink-has-double-slash.
The test for it failed as a result of commit 560482dd.

This commit fixes the failure in the test suite, but leaves the
resolution potentially unsolved.  The matter will be resolved properly
with Moo triggers in the near future.

Gbp-Dch: ignore