Do not attempt to decode UTF-8 in hashbang interpreters in scripts.
Prior to removal, the decoding failed in the two source packages
magyarispell_1.6.1-2.dsc and ldc_1.24.0-1.dsc, but their maintainers
can do nothing about it. A tag is therefore useless.

For shipped files, the non UTF-8 encoding will be caught elsewhere.

A conversion of the interpreter string should instead be attempted,
provided it is needed at all, in the check verifying that an
interpreter is well known. That check runs only on shipped files, and
would not flag these cases about which, again, a maintainer can do
nothing.

Several related issues will be discussed in a message scheduled to be
posted to debian-lint-maint after this change is committed.