Load and cache Lintian::Data objects in profiles; drop their global data structures.
Now the data structures are no longer global. They also no longer rely
on a having a single, global profile. The checks can access copies and
keep them in locally scoped variables or in Moo attributes.

This change will likely break libconfig-model-dpkg-perl, the only
package known to rely on Lintian's proprietary data structures. For
now, please load the data with:

    my $data = $profile->load_data(...);

instead of

    my $data = Lintian::Data->new(...);

That is a temporary fix; the interface is scheduled to change again in
the near future.

The use of Lintian's data objects by third parties is expected to stop
with the resolution of Bug#968011. That step is necessary to develop
Lintian safely on a system that has Lintian installed, and also run
autopkgtests on such a system.

It would be helpful if libconfig-model-dpkg-perl could please find a
way to obtain the policy release dates directly from the responsible
team, as requested in Bug#968000. Thank you!