In .perlcriticrc, use long names in config blocks instead of short names in the 'include' line.
This technique of enabling policies under the whitelist regime
('only') makes the file easier to understand and manage.

Negative config blocks were added as documentation for all policies
that were mentioned but not enabled. The only exceptions are the short
names UselessInitialization and NullStatements, which trigger errors
in their long forms.

They are missing because the package libperl-critic-pulp-perl was
removed in commit 3b65b36e. We hope to add those policies back at a
later date.

The perlcritic config file does not allow continuation lines, which
made the 'include' line too long.