Disregard the perlcritic policy Subroutines::ProhibitExplicitReturnUndef.
This policy is part of a broader effort to eliminate the NULL from
Lintian. [1] Here, it conflicts with the policy Freenode::EmptyReturn,
which protect against unspecified return values when one is expected.

The latter serves a more immediate need. The effort to eradicate the
NULL from Lintian will be revisited.

[1] https://www.infoq.com/presentations/Null-References-The-Billion-Dollar-Mistake-Tony-Hoare/