Allow tilde in names of installed JAR archives. (Closes: #976681)
What an interesting little bug! The regular expression that validates
the names of installed JAR files is not widely used. It was incorrect.
A tilde was added to the list of acceptable characters.

The three false positives mentioned in the bug report were cured. The
tag bad-jar-name disappeared from each one:

$ bin/lintian /mirror/debian/pool/main/libs/libsis-base-java/libsis-base-java_18.09\~pre1+git20180928.45fbd31+dfsg-1_all.deb
[produces no output]

$ bin/lintian /mirror/debian/pool/main/libj/libjsr305-java/libjsr305-java-doc_0.1\~+svn49-10_all.deb
E: libjsr305-java-doc: malformed-override Unknown tag no-upstream-changelog in line 1

The package with the override no longer used it, and said so:

$ bin/lintian /mirror/debian/pool/main/libp/libpj-java/libpj-java_0.0\~20150107+dfsg-4_all.deb
I: libpj-java: unused-override bad-jar-name usr/share/java/pj-*
P: libpj-java: repeated-path-segment pj usr/share/maven-repo/edu/rit/pj/pj/

Thanks to Pierre Gruet for bringing the matter to our attention. The
filing was so well-researched, it was a pleasure to solve!