Use PKGVERSION_REGEX to allow tilde in installed JAR names; add bg nuber references. (See: #976681)
Upon reflection, the bug was a better solved in the manner presented
here. It reverts the change made in commit 48774d01. Lintian's
behavior with respect to the three packages mentioned there, however,
remained unchanged.

Employs a second, existing regular expression to accommodate the tilde
in the file name. Its application here was not immediately obvious
because, unlike elsewhere, name and version were not separated by an
underscore but by a hyphen.

Gbp-Dch: ignore