Update manual references for the first time in almost two years.
Finally propagates Debian's adption of version 3.0 of the Filesystem
Hierarchy Standard from 2018, as implemented by the resolution of
Bug#787816.

Given the relative timing of events, it is unclear why the last two
updates of the manual references with commits 0e40beba (January 2019)
and commit 063caa42 (January of 2020) did not pick up the change to
FHS 3.0.