Use current program names in help messages and version announcements. (Closes: #977261)
Some program names changed multiple times recently, but were hardcoded
(instead of using $0) in order to accommodate the symbolic link for
the even more outdated and ambiguous name 'lintian-info'.

It left help messages and version announcements in inconsistent
states. which was confusing to users. Fixed here.

Thanks to Robert Luberda for bringing the matter to our attention!