Convert possible-gpl-code-linked-with-openssl to a classification tag. (Closes: #972181)
The tag was converted to a classification tag. In that capacity may
help with research into adoption rates of OpenSSL in Debian since the
barriers to its more widespread use have disappeared.

The tag will now enter our normal cleanup and repurposing efforts.  It
may be dropped completely if there is no residual utility.

Classification tags are not marked in package runs on our website.
They are also invisible to users of our command-line utilities (unless
special options are enabled).

For all practical purposes, the tag is gone.