Add '::Check' to all Perl module names for checks.
A grouping for these modules in Perl's name space is a precursor to
ensure that the correct checks are being loaded, and that all checks
requested are also being run.

In addition, much of Perl's internal and external tooling works less
well with non-standard package names.

Together with the changes to follow, this may break pkg-js-tools and
pkg-perl-tools. The maintainers of those packages will be advised
separately, as has been our custom.