Capitalize module names for checks in camel case; drop underscores.
Lintian tried for a long time to make the module names for checks
similar to their names. It is not possible to make them the same.
(Module names may not contain hyphens; not sure about dots.) This
commit chooses an alternative naming scheme.

The new names are more in line with Perl customs. We introduce them
here in preparation to move the modules to their correct location in
our name space. Symbolic links will be provided as a convenience where
appropriate.

The move will prevent modules from being absent due to typos. It will
also help to ensure that all modules are loaded. Some of these
assertions did not appear possible without the perlcritic policy
Freenode::PackageMatchesFilename, which we hope to enable shortly.

Aside from the broader name tranlations, this commit currently
special-cases the dot in the check 'init.d' in two conversions from
camel case in which otherwise a hyphen would be inserted.

Together with the changes to follow, this may break pkg-js-tools and
pkg-perl-tools. The maintainers of those packages will be advised
separately, as has been our custom.