Ignore SIGWINCH in the test suite. (Closes: #974575)
The source of this odd behavior was not identified. It continued under
MCE::Loop when that was used experimentally instead of IO::Async.
Based on past experience, there was a concern about IO::Async and
SIGCHLD.  Either module, however, helped only briefly to prepare
tests.

The problem occurred already in the test scripts before any custom
loops were started, so the loop modules are not at fault.

The bulk of the work in private/runtests took place under the aegis of
TAP::Harness->aggregate_tests.  As noted in the bug, something in the
module initialization caused it, and probably not in Perl.

The upcoming merging of Check::Info into Profile and Group appeared to
cause a similar problem with SIGCHLD. That issue, and its mitigation,
are described here:

    https://stackoverflow.com/a/60761593

A proper identification and a solution remain outstanding.