Switch to MCE::Loop for test prparation in ./private/runtests.
This is an experimental change since MCE performed relatively well to
generate tags for the website.

It is not a vote against IO::Async. We are simply desparate to get to
the bottom of some signal handling issues in TAP::Harness that are
probably caused by something else. For details, please see Bug#974575.

Gbp-Dch: ignore