Move all checks into the module name space under ./lib; provide symbolic link.
Traditionally, the module names for checks in Lintian did not match
their location in the file system. In Perl, that leads to several
serious shortcomings that are remedied in this commit series.

Here, checks are moved to lib/Lintian/Check. A symbolic link will be
provided as a convenience in the next commit.

The move prevents modules from being absent due to typos. It also
ensures that all modules are loaded under the names we expect. Some of
these assertions are not a possible without the help of the perlcritic
policy Freenode::PackageMatchesFilename, which we hope to enable
shortly.

Together with the changes to follow, this may break pkg-js-tools and
pkg-perl-tools. The maintainers of those packages will be advised
separately, as has been our custom.