For manual pages, extract section information correctly from .TH header. (Closes: #962601)
The value of the original filing of this bug is unclear.  The use of a
comma-separated list with whitespaces that separate multiple names is
very rare.

The manual page mentioned in the bug report did not do so when this
message was written. It is unclear where the filer found it.  At the
time of filing, the version then in unstable (2.6.9-1) shipped a
manual page that read:

    .TH MOUNT.CIFS 8 "" "" ""

The version now in unstable (2.6.11-10 ships a header with whitespace
but no comma. That is likely an error:

    .TH MOUNT.CIFS MOUNT.SMB3 8 "" "" ""

The only manual page that was remotely relevant that we were able to
find was gv2gml.1 from graphviz, which had a comma but no whitespace.
Thanks to DLange fo pointing it out! That header read:

    .TH GML2GV,GV2GML 1 "24 June 2011"

Either way, we hope that this change will put the bug to rest. At a
minimum, it seems to silence the run time errors that have been
showing up since we attempted to resolve this bug the first time
around.